Write the API relaying middle server
Take the path part of the URL that the middle server receives and
send it to the remote server. Relay the response back to
the initial client via the middle server.

Example: If the client visits the middle server via the URL
http://localhost/restaurant then give the client the HTTP response
of https://menu.dckube.scilifelab.se/api/restaurant via the middle server.